requirements.txt
click
pandas
Pillow
numpy
requests
protobuf
